Add error handling. Move interface to appEntity file.
