+++ src/private/interfaces.ts
@@ -159,36 +159,3 @@ export interface UserJoinedTeamsInformation {
    */
   userJoinedTeams: TeamInformation[];
 }
-
-/**
- * @private
- * Hide from docs
- * --------
- * Information on an app entity
- */
-export interface AppEntityConfiguration {
-  /**
-   * App ID of the application
-   */
-  appId: string;
-
-  /**
-   * URL for the application's icon
-   */
-  appIconUrl: string;
-
-  /**
-   * Content URL for the app entity
-   */
-  contentUrl: string;
-
-  /**
-   * The display name for the app entity
-   */
-  displayName: string;
-
-  /**
-   * Website URL for the app entity. It is meant to be opened by the user in a browser.
-   */
-  websiteUrl: string;
-}
